FIX imports for renaming auth module
